esp32-ocr-server/requirements.txt
Flask
pytesseract
Pillow
gTTS
